requirements.txt
boto3>=1.34.0
pyyaml>=6.0
requests>=2.31.0
tenacity>=8.2.0
redis>=5.0.0
